Enable VR launching without eye tracking. Separated run logic into CellTrackingBase
